fix: split TestCascadeNewEntries into test_test_filter_cascade.py to satisfy 1000-line limit
